API CHANGE! add series parent reference to episode
This is a major API change as episode structures now contain a reference
to their parent series.
Moving to the new API is straight forward. Some Episode functions take
an extra "Series" argument, and instead of the series-id a full series
structure must be passed. Check the documentation for details.