version bump: alpha 0.1 release
this is the first release. the API should be stable now, at the very
least until I add an async API (and even then I plan to keep the sync
API as is).
I couldn't find any major bug to this point, so I think this should make
a rather sold first alpha release.
For the release etvdb_cli has been removed and is developed externally
now, as it will get more dependencies that are not necessary for etvdb.